test ok

+++ test.cpp
@@ -18,9 +18,9 @@ using namespace std;
 void test_multiples(vector<est_premier> funcs, bigint a, bigint b, int n)
 {
 
-    vector<long int> nbs;
+    vector<bigint> nbs;
     std::default_random_engine re(42);
-    std::uniform_int_distribution<long int> distrib {(long int)a,(long int)b};
+    std::uniform_int_distribution<bigint> distrib {a,b};
     for(int i = 0; i < n; i++)
     {
         nbs.push_back(distrib(re));
